Moved the filesystem methods to a new file
