+++ make_includes/obj_000.inc
@@ -2,10 +2,10 @@
 # object files (generic 000)
 ##########################################################################
 
-src/funcs.o: src/funcs.c src/iGame_cat.h src/strfuncs.h
+src/funcs.o: src/funcs.c src/iGame_cat.h src/strfuncs.h src/fsfuncs.h
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/funcs.c
 
-src/iGameGUI.o: src/iGameGUI.c src/iGameGUI.h src/version.h src/iGame_cat.h
+src/iGameGUI.o: src/iGameGUI.c src/iGameGUI.h src/iGame_cat.h src/fsfuncs.h
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/iGameGUI.c
 
 src/iGameMain.o: src/iGameMain.c
@@ -13,6 +13,9 @@ src/iGameMain.o: src/iGameMain.c
 
 src/strfuncs.o: src/strfuncs.c src/strfuncs.h
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/strfuncs.c
+
+src/fsfuncs.o: src/fsfuncs.c src/fsfuncs.h
+	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/fsfuncs.c
 
 src/iGame_cat.o: src/iGame_cat.c
 	$(CC) $(CFLAGS) $(INCLUDES) -o $@ src/iGame_cat.c
