+++ src/iGameExtern.h
@@ -149,7 +149,7 @@ void settings_use();
 const unsigned char* GetMBString(const unsigned char* ref);
 void joy_left();
 void joy_right();
-void open_current_dir();
+// void open_current_dir();
 ULONG get_wb_version();
 
 #endif
